refactor(T6): move manifest-grouping into reports feature dir
bucketLinesBySponsor had exactly one production caller
(LinesManifestView). Moving the helper and its test out of
src/utils/ and co-locating them with the feature removes the
one-off-file-in-shared-utils anti-pattern flagged in PR review.
