+++ src/components/sponsors/reports/__tests__/manifest-grouping.test.js
