+++ src/components/sponsors/reports/manifest-grouping.js
